pyproject.toml
[project]
name = "redmince-fastmcp"
version = "0.1.0"
description = "MCP server bridging AI agents to Redmine via OAuth 2.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.0.0,<4.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
